apexes do not use apex_name
Since apex'es name is determined by the manifest json, no need to set it
in Android.bp.

When apex_name is set, symbol files of apex are installed based on it,
which caused build errors when there's multi-install apexes. (Two apexes
with the same apex_name). To work around the error,
multi_install_skip_symbol_files property was introduced.

This change also cleans up the workarounds as well.

Bug: 267581665
Test: aosp_cf_x86_64_phone (m & launch)
Change-Id: I3eb438cd5913a80fd08b59afe4aac76dea0694bb
Signed-off-by: spezi77 <spezi7713@gmx.net>